requirements.txt
azure-functions>=1.18.0
azure-monitor-ingestion>=1.0.0
azure-identity>=1.15.0
requests>=2.31.0
cryptography>=41.0.0
cffi>=1.15.0
